requirements.txt
streamlit
together
pandas
matplotlib
python-docx
openpyxl
pymupdf
easyocr
pillow
